scalaPB implementation moved to specialized project
