+++ core-scalapb/src/main/scala/com/avast/grpc/jsonbridge/scalapb/ScalaPBReflectionGrpcJsonBridge.scala
@@ -1,0 +1,5 @@
+package com.avast.grpc.jsonbridge.scalapb
+
+import com.avast.grpc.jsonbridge.ReflectionGrpcJsonBridge
+
+object ScalaPBReflectionGrpcJsonBridge extends ReflectionGrpcJsonBridge(ScalaPBServiceHandlers)
