+++ core-scalapb/src/test/protobuf/TestServices.proto
