+++ core-scalapb/src/test/scala/com/avast/grpc/jsonbridge/scalapbtest/TestServiceImpl.scala
